move account generator items to new page
