+++ src/constants/slug.js
@@ -2,6 +2,7 @@
 // These are used in to build the url and are the "page name"
 const SLUG = {
   HOME: '',
+  ACCOUNT_CREATOR: 'account-creator',
   EXPLORER: 'explorer',
   TXBUILDER: 'txbuilder',
   TXSIGNER: 'txsigner',
