+++ src/reducers/root.js
@@ -1,5 +1,5 @@
 import {combineReducers} from 'redux';
-import introSetup from './introSetup';
+import accountCreator from './accountCreator';
 import endpointExplorer from './endpointExplorer';
 import transactionBuilder from './transactionBuilder';
 import transactionSigner from './transactionSigner';
@@ -8,7 +8,7 @@ import network from './network';
 import routing from './routing';
 
 const rootReducer = combineReducers({
-  introSetup,
+  accountCreator,
   endpointExplorer,
   transactionBuilder,
   transactionSigner,
