+++ src/styles/_AccountCreator.scss
@@ -1,0 +1,18 @@
+.AccountCreator {
+}
+  .AccountCreator__container {
+    max-width: 35em;
+  }
+  .AccountCreator__lead {
+    margin-top: 1em;
+  }
+  .AccountCreator__section {
+    padding-top: 2em;
+    padding-bottom: 2em;
+  }
+  .AccountCreator__separator {
+    border-bottom: 1px solid $s-color-neutral6;
+  }
+  .AccountCreator__spaceTop {
+    margin-top: 1em;
+  }
