+++ src/styles/_Introduction.scss
@@ -8,11 +8,3 @@
   .Introduction__lead {
     margin-top: 1em;
   }
-  .Introduction__section {
-    padding-top: 2em;
-    padding-bottom: 2em;
-    border-top: 1px solid $s-color-neutral6;
-  }
-  .Introduction__spaceTop {
-    margin-top: 1em;
-  }
