+++ src/styles/main.scss
@@ -28,6 +28,8 @@
 
 @import 'Introduction';
 
+@import 'AccountCreator';
+
 @import 'EndpointExplorer';
   @import 'EndpointPicker';
   @import 'EndpointSetup';
